jailbreak_detection/requirements.txt
streamlit==1.32.0
Pillow==10.2.0
numpy==1.26.4
opencv-python==4.9.0.80
scikit-learn==1.4.0
tqdm==4.66.1